requirements.txt
streamlit>=1.33
pandas>=2.2
numpy>=1.26
scikit-learn>=1.4
joblib>=1.3
python-dotenv>=1.0
plotly>=5.20
requests>=2.31
